1.3.6

+++ core/src/main/java/me/athlaeos/valhallammo/crafting/recipetypes/DynamicCookingRecipe.java
@@ -31,7 +31,7 @@ public class DynamicCookingRecipe implements ValhallaRecipe, ValhallaKeyedRecipe
     private boolean hiddenFromBook = false;
     private boolean requireValhallaTools = false;
     private List<DynamicItemModifier> modifiers = new ArrayList<>();
-    private int experience = 1;
+    private float experience = 1;
     private Collection<String> validations = new HashSet<>();
     private boolean unlockedForEveryone = false;
     private final CookingRecipeType type;
@@ -63,7 +63,7 @@ public class DynamicCookingRecipe implements ValhallaRecipe, ValhallaKeyedRecipe
     public int getCookTime() { return cookTime; }
     public boolean requireValhallaTools() { return requireValhallaTools; }
     public List<DynamicItemModifier> getModifiers() { return modifiers; }
-    public int getExperience() { return experience; }
+    public float getExperience() { return experience; }
     public Collection<String> getValidations() {return validations;}
     public boolean isUnlockedForEveryone() { return unlockedForEveryone; }
     public CookingRecipeType getType() { return type; }
@@ -82,7 +82,7 @@ public class DynamicCookingRecipe implements ValhallaRecipe, ValhallaKeyedRecipe
         this.modifiers = modifiers;
         DynamicItemModifier.sortModifiers(this.modifiers);
     }
-    public void setExperience(int experience) { this.experience = experience; }
+    public void setExperience(float experience) { this.experience = experience; }
     public void setValidations(Collection<String> validations) {this.validations = validations;}
     public void setUnlockedForEveryone(boolean unlockedForEveryone) { this.unlockedForEveryone = unlockedForEveryone; }
     public void setHiddenFromBook(boolean hiddenFromBook) { this.hiddenFromBook = hiddenFromBook; }
